requirements.txt
numpy>=1.19
scipy>=1.6
theano-pymc>=1.1
pymc3>=3.11